// frontend/src/app/api/model/sale-create-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SaleItemRequest } from './sale-item-request';


export interface SaleCreateDto { 
    customerId?: number;
    items: Array<SaleItemRequest>;
}